backend/requirements.txt
Flask
Flask-CORS
python-dotenv
requests
twilio
redis
gunicorn
psycopg2-binary
